Refactor Level, Orb (WIP)

+++ desktop/src/com/jmolina/orb/desktop/DesktopLauncher.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.Files;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.jmolina.orb.OrbApp;
-import com.jmolina.orb.assets.Asset;
+import com.jmolina.orb.var.Asset;
 import com.jmolina.orb.var.Var;
 
 public class DesktopLauncher {
@@ -12,8 +12,8 @@ public class DesktopLauncher {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
 
 		config.fullscreen = false;
-		config.height = Var.SCREEN_HEIGHT;
-		config.width = Var.SCREEN_WIDTH;
+		config.height = (int) Var.SCREEN_HEIGHT;
+		config.width = (int) Var.SCREEN_WIDTH;
 		config.title = Var.APP_NAME + " " + Var.APP_VERSION;
 		config.addIcon(Asset.APP_ICON_128, Files.FileType.Internal);
 		config.addIcon(Asset.APP_ICON_64, Files.FileType.Internal);
